# swagger.json
{"openapi":"3.0.0","paths":{"/ping":{"get":{"operationId":"AppController_ping","summary":"Consulta status do serviço","description":"Rota utilizada para consultar o status do sistema e de suas\n    dependências.","parameters":[],"responses":{"200":{"description":"Serviço online."},"503":{"description":"Serviço offline ou instável","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PingResponse"}}}}},"tags":["App"]}},"/users":{"post":{"operationId":"UsersController_create","summary":"Criação de usuários","description":"Rota para criação de usuários do sistema.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"Retorna o usuário criado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultSingleResponse"}}}},"400":{"description":"Erro de validação de dados.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultErrorResponse"}}}},"500":{"description":"Erro interno"}},"tags":["Usuários"]},"get":{"operationId":"UsersController_findAll","summary":"Consulta de usuários","description":"Rota para consulta de usuários de forma paginada,\n    para que seja de melhor consumo através do frontend.","parameters":[{"name":"page","required":false,"in":"query","description":"Numero da pagina a ser visualizada","schema":{"default":1,"type":"number"}}],"responses":{"200":{"description":"Retorna os dados de usuários requisitados.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultPaginatedResponse"}}}},"401":{"description":"Usuário não autorizado."},"500":{"description":"Erro interno"}},"tags":["Usuários"]}},"/users/{id}":{"get":{"operationId":"UsersController_findOne","summary":"Consulta de usuário","description":"Rota para consulta de um único usuário.","parameters":[{"name":"id","required":true,"in":"path","description":"ID do usuário a ser consultado","schema":{"type":"string"}}],"responses":{"200":{"description":"Retorna o usuário solicitado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultSingleResponse"}}}},"401":{"description":"Usuário não autorizado."},"500":{"description":"Erro interno"}},"tags":["Usuários"]},"patch":{"operationId":"UsersController_update","summary":"Alteração de dados do usuário","description":"Rota para alteração de dados do usuário.","parameters":[{"name":"id","required":true,"in":"path","description":"ID do usuário a ser alterado","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"Retorna o usuário alterado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultSingleResponse"}}}},"400":{"description":"Erro de validação de dados.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultErrorResponse"}}}},"401":{"description":"Usuário não autorizado."},"500":{"description":"Erro interno"}},"tags":["Usuários"]},"delete":{"operationId":"UsersController_remove","summary":"Desabilitação de usuário","description":"Rota para desabilitar o usuário.","parameters":[{"name":"id","required":true,"in":"path","description":"ID do usuário a ser desabilitado","schema":{"type":"string"}}],"responses":{"204":{"description":"Usuário desabilitado com sucesso."},"401":{"description":"Usuário não autorizado."},"500":{"description":"Erro interno"}},"tags":["Usuários"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"responses":{"201":{"description":""}}}},"/company":{"post":{"operationId":"CompanyController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCompanyDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"CompanyController_findAll","parameters":[{"name":"page","required":false,"in":"query","description":"Numero da pagina a ser visualizada","schema":{"default":1,"type":"number"}}],"responses":{"200":{"description":""}}}},"/company/{id}":{"get":{"operationId":"CompanyController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"CompanyController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCompanyDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"CompanyController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}}},"info":{"title":"Controla fácil","description":"Este projeto tem por objetivo centralizar as principais requisições que serão enviadas ao sistema controla fácil, que é um sistema focado em controle de estoque para prestadores de serviço.","version":"1.0","contact":{"name":"Rafael P. Scholant","url":"https://www.linkedin.com/in/rafaelscholant","email":"rafael.scholant@gmail.com"}},"tags":[{"name":"App","description":"Rotas de gerenciamento do serviço."}],"servers":[{"url":"http://localhost:3000","description":"Ambiente de desenvolvimento"},{"url":"https://controla-facil.herokuapp.com","description":"Ambiente de produção"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http","description":"Politica de autenticação de usuários padrão do sistema"}},"schemas":{"PingResponse":{"type":"object","properties":{"message":{"type":"string","readOnly":true,"example":"pong"},"health":{"type":"object","readOnly":true,"example":{"status":"up","info":{"database":{"status":"up"}},"error":{},"details":{}}}},"required":["message","health"]},"CreateUserDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"companyId":{"type":"string"}},"required":["name","email","password","companyId"]},"DefaultSingleResponse":{"type":"object","properties":{"status":{"type":"number","readOnly":true,"example":201},"message":{"type":"string","readOnly":true,"example":"Mensagem exemplo"},"data":{"type":"object","readOnly":true,"example":{}}},"required":["status","message","data"]},"DefaultErrorResponse":{"type":"object","properties":{"statusCode":{"type":"number","readOnly":true,"example":400},"message":{"type":"string","readOnly":true,"example":"Mensagem exemplo"},"data":{"type":"object","readOnly":true,"example":{}}},"required":["statusCode","message","data"]},"DefaultPaginatedResponse":{"type":"object","properties":{"data":{"readOnly":true,"example":[{}],"type":"array","items":{"type":"string"}},"additional":{"type":"object","readOnly":true,"example":{}},"page":{"type":"number","readOnly":true,"example":1},"count":{"type":"number","readOnly":true,"example":1}},"required":["data","additional","page","count"]},"UpdateUserDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"companyId":{"type":"string"}}},"CreateCompanyDto":{"type":"object","properties":{"name":{"type":"string"},"socialName":{"type":"string"},"cnpj":{"type":"string"}},"required":["name","socialName","cnpj"]},"UpdateCompanyDto":{"type":"object","properties":{"name":{"type":"string"},"socialName":{"type":"string"},"cnpj":{"type":"string"}}}}}}